multiplicacion

+++ Clases/src/mx/rafex/cursos/introduccion/clases/ejercicios/Operacion.java
@@ -26,7 +26,7 @@ public class Operacion {
 
     public static Double multiplicar(Double... numeros) {
 
-        Double resultado = new Double(0);
+        Double resultado = new Double(1);
 
         for (final Double numero : numeros)
             resultado *= numero;
